Move tests to correct test file
Seems like a common error was to write tests in this file,
because it was named `test_create_model.py`.
