+++ tests/test_create_model.py
@@ -1,6 +1,5 @@
-import platform
 import re
-from typing import Generic, Optional, Tuple, TypeVar
+from typing import Generic, Tuple, TypeVar
 
 import pytest
 from typing_extensions import Annotated
@@ -18,7 +17,6 @@ from pydantic import (
     field_validator,
     validator,
 )
-from pydantic.fields import ModelPrivateAttr
 
 
 def test_create_model() -> None:
@@ -319,203 +317,6 @@ def test_create_model_with_doc():
     model = create_model('FooModel', foo=(str, ...), bar=(int, 123), __doc__='The Foo model')
     assert model.__name__ == 'FooModel'
     assert model.__doc__ == 'The Foo model'
-
-
-@pytest.mark.parametrize('base', [ModelPrivateAttr, object])
-@pytest.mark.parametrize('use_annotation', [True, False])
-def test_private_descriptors(base, use_annotation):
-    set_name_calls = []
-    get_calls = []
-    set_calls = []
-    delete_calls = []
-
-    class MyDescriptor(base):
-        def __init__(self, fn):
-            super().__init__()
-            self.fn = fn
-            self.name = ''
-
-        def __set_name__(self, owner, name):
-            set_name_calls.append((owner, name))
-            self.name = name
-
-        def __get__(self, obj, type=None):
-            get_calls.append((obj, type))
-            return self.fn(obj) if obj else self
-
-        def __set__(self, obj, value):
-            set_calls.append((obj, value))
-            self.fn = lambda obj: value
-
-        def __delete__(self, obj):
-            delete_calls.append(obj)
-
-            def fail(obj):
-                # I have purposely not used the exact formatting you'd get if the attribute wasn't defined,
-                # to make it clear this function is being called, while also having sensible behavior
-                raise AttributeError(f'{self.name!r} is not defined on {obj!r}')
-
-            self.fn = fail
-
-    class A(BaseModel):
-        x: int
-
-        if use_annotation:
-            _some_func: MyDescriptor = MyDescriptor(lambda self: self.x)
-        else:
-            _some_func = MyDescriptor(lambda self: self.x)
-
-        @property
-        def _double_x(self):
-            return self.x * 2
-
-    assert set(A.__private_attributes__) == {'_some_func'}
-    assert set_name_calls == [(A, '_some_func')]
-
-    a = A(x=2)
-
-    assert a._double_x == 4  # Ensure properties with leading underscores work fine and don't become private attributes
-
-    assert get_calls == []
-    assert a._some_func == 2
-    assert get_calls == [(a, A)]
-
-    assert set_calls == []
-    a._some_func = 3
-    assert set_calls == [(a, 3)]
-
-    assert a._some_func == 3
-    assert get_calls == [(a, A), (a, A)]
-
-    assert delete_calls == []
-    del a._some_func
-    assert delete_calls == [a]
-
-    with pytest.raises(AttributeError, match=r"'_some_func' is not defined on A\(x=2\)"):
-        a._some_func
-    assert get_calls == [(a, A), (a, A), (a, A)]
-
-
-def test_private_attr_set_name():
-    class SetNameInt(int):
-        _owner_attr_name: Optional[str] = None
-
-        def __set_name__(self, owner, name):
-            self._owner_attr_name = f'{owner.__name__}.{name}'
-
-    _private_attr_default = SetNameInt(1)
-
-    class Model(BaseModel):
-        _private_attr_1: int = PrivateAttr(default=_private_attr_default)
-        _private_attr_2: SetNameInt = SetNameInt(2)
-
-    assert _private_attr_default._owner_attr_name == 'Model._private_attr_1'
-
-    m = Model()
-    assert m._private_attr_1 == 1
-    assert m._private_attr_1._owner_attr_name == 'Model._private_attr_1'
-    assert m._private_attr_2 == 2
-    assert m._private_attr_2._owner_attr_name == 'Model._private_attr_2'
-
-
-def test_private_attr_default_descriptor_attribute_error():
-    class SetNameInt(int):
-        def __get__(self, obj, cls):
-            return self
-
-    _private_attr_default = SetNameInt(1)
-
-    class Model(BaseModel):
-        _private_attr: int = PrivateAttr(default=_private_attr_default)
-
-    assert Model.__private_attributes__['_private_attr'].__get__(None, Model) == _private_attr_default
-
-    with pytest.raises(AttributeError, match="'ModelPrivateAttr' object has no attribute 'some_attr'"):
-        Model.__private_attributes__['_private_attr'].some_attr
-
-
-def test_private_attr_set_name_do_not_crash_if_not_callable():
-    class SetNameInt(int):
-        __set_name__ = None
-
-    _private_attr_default = SetNameInt(2)
-
-    class Model(BaseModel):
-        _private_attr: int = PrivateAttr(default=_private_attr_default)
-
-    # Checks below are just to ensure that everything is the same as in `test_private_attr_set_name`
-    # The main check is that model class definition above doesn't crash
-    assert Model()._private_attr == 2
-
-
-def test_del_model_attr():
-    class Model(BaseModel):
-        some_field: str
-
-    m = Model(some_field='value')
-    assert hasattr(m, 'some_field')
-
-    del m.some_field
-
-    assert not hasattr(m, 'some_field')
-
-
-@pytest.mark.skipif(
-    platform.python_implementation() == 'PyPy',
-    reason='In this single case `del` behaves weird on pypy',
-)
-def test_del_model_attr_error():
-    class Model(BaseModel):
-        some_field: str
-
-    m = Model(some_field='value')
-    assert not hasattr(m, 'other_field')
-
-    with pytest.raises(AttributeError, match='other_field'):
-        del m.other_field
-
-
-def test_del_model_attr_with_privat_attrs():
-    class Model(BaseModel):
-        _private_attr: int = PrivateAttr(default=1)
-        some_field: str
-
-    m = Model(some_field='value')
-    assert hasattr(m, 'some_field')
-
-    del m.some_field
-
-    assert not hasattr(m, 'some_field')
-
-
-@pytest.mark.skipif(
-    platform.python_implementation() == 'PyPy',
-    reason='In this single case `del` behaves weird on pypy',
-)
-def test_del_model_attr_with_privat_attrs_error():
-    class Model(BaseModel):
-        _private_attr: int = PrivateAttr(default=1)
-        some_field: str
-
-    m = Model(some_field='value')
-    assert not hasattr(m, 'other_field')
-
-    with pytest.raises(AttributeError, match="'Model' object has no attribute 'other_field'"):
-        del m.other_field
-
-
-def test_del_model_attr_with_privat_attrs_twice_error():
-    class Model(BaseModel):
-        _private_attr: int = 1
-        some_field: str
-
-    m = Model(some_field='value')
-    assert hasattr(m, '_private_attr')
-
-    del m._private_attr
-
-    with pytest.raises(AttributeError, match="'Model' object has no attribute '_private_attr'"):
-        del m._private_attr
 
 
 def test_create_model_protected_namespace_default():
